# rest-service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: This is the specification of the required API endpoints for the SWKOM
    server.
  title: SWKOM
  version: 1.0.11
servers:
- url: https://localhost:8081
tags:
- description: Documents user uploads and maintains
  name: documents
paths:
  /documents:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Document'
                type: array
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/Document'
                type: array
          description: A list of documents is returned.
        "401":
          description: Access token is missing or invalid
        "404":
          description: No documents found.
      security:
      - swkom_auth:
        - read:documents
      summary: Get all documents
      tags:
      - documents
      x-accepts:
      - application/json
      - application/xml
      x-tags:
      - tag: documents
    post:
      description: Allows users to create a new document.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: Document successfully created
        "400":
          description: Invalid input
        "422":
          description: Validation exception
      security:
      - swkom_auth: []
      summary: Creates a new document.
      tags:
      - documents
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: documents
    put:
      description: Allows users to update a document.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: Document successfully created
        "400":
          description: Invalid input
        "422":
          description: Validation exception
      security:
      - swkom_auth: []
      summary: Updates document.
      tags:
      - documents
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: documents
  /documents/{documentId}:
    delete:
      description: Deletes an existing document. Only allowed if the user owns the
        document.
      parameters:
      - description: ID of the document to retrieve.
        explode: false
        in: path
        name: documentId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          description: Document successfully deleted
        "401":
          description: Access token is missing or invalid
        "404":
          description: Document not found
        "409":
          description: Conflict in document deletion
      security:
      - swkom_auth: []
      summary: Deletes an existing document
      tags:
      - documents
      x-accepts:
      - application/json
      x-tags:
      - tag: documents
    get:
      parameters:
      - description: ID of the document to retrieve.
        explode: false
        in: path
        name: documentId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
            application/xml:
              schema:
                $ref: '#/components/schemas/Document'
          description: Document retrieved successfully
        "401":
          description: Access token is missing or invalid
        "404":
          description: Document not found
      security:
      - swkom_auth:
        - read:documents
      summary: Retrieve an existing document
      tags:
      - documents
      x-accepts:
      - application/json
      - application/xml
      x-tags:
      - tag: documents
  /documents/{documentId}/content:
    get:
      parameters:
      - explode: false
        in: path
        name: documentId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentContent'
            application/xml:
              schema:
                $ref: '#/components/schemas/DocumentContent'
          description: Document retrieved successfully
        "401":
          description: Access token is missing or invalid
        "404":
          description: Document not found
      security:
      - swkom_auth:
        - read:documents
      tags:
      - documents
      x-accepts:
      - application/json
      - application/xml
      x-tags:
      - tag: documents
    put:
      description: Allows users to upload new content for an existing document.
      parameters:
      - explode: false
        in: path
        name: documentId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: Document content successfully updated
        "400":
          description: Invalid input
        "422":
          description: Validation exception
      security:
      - swkom_auth:
        - write:documents
        - read:documents
      summary: Updates the content of an existing document.
      tags:
      - documents
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: documents
components:
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
  schemas:
    Document:
      example:
        uploaded_date: uploaded_date
        description: This is a description of the doc
        id: 550e8400-e29b-41d4-a716-446655440000
        username: name
      properties:
        id:
          example: 550e8400-e29b-41d4-a716-446655440000
          format: uuid
          type: string
        username:
          example: name
          type: string
        description:
          example: This is a description of the doc
          type: string
        uploaded_date:
          type: string
      type: object
      xml:
        name: document
    DocumentContent:
      example:
        id: 550e8400-e29b-41d4-a716-446655440000
        content: content
      properties:
        id:
          example: 550e8400-e29b-41d4-a716-446655440000
          format: uuid
          type: string
        content:
          type: string
      type: object
  securitySchemes:
    swkom_auth:
      scheme: bearer
      type: http